+++ typescript/deploy-sdk/eslint.config.mjs
@@ -1,9 +1,8 @@
-import { jsRules, typescriptRules } from '@hyperlane-xyz/eslint-config';
+import { defaultConfig } from '@hyperlane-xyz/eslint-config';
 
 export default [
   { ignores: ['dist/**'] },
-  ...jsRules,
-  ...typescriptRules,
+  ...defaultConfig,
   {
     name: 'deploy-sdk-overrides',
     rules: {