+++ typescript/svm-sdk/src/clients/protocol.ts
@@ -18,6 +18,8 @@ import type { IRawWarpArtifactManager } from '@hyperlane-xyz/provider-sdk/warp';
 import { assert } from '@hyperlane-xyz/utils';
 import { address as parseAddress } from '@solana/kit';
 
+import { type IRawMailboxArtifactManager } from '@hyperlane-xyz/provider-sdk/mailbox';
+import { type IRawValidatorAnnounceArtifactManager } from '@hyperlane-xyz/provider-sdk/validator-announce';
 import { SvmHookArtifactManager } from '../hook/hook-artifact-manager.js';
 import { SvmIsmArtifactManager } from '../ism/ism-artifact-manager.js';
 import { createRpc } from '../rpc.js';
@@ -70,6 +72,20 @@ export class SvmProtocolProvider implements ProtocolProvider {
   ): IRawWarpArtifactManager {
     const rpc = createRpc(this.getRpcUrls(chainMetadata)[0]);
     return new SvmWarpArtifactManager(rpc);
+  }
+
+  createMailboxArtifactManager(
+    _chainMetadata: ChainMetadataForAltVM,
+  ): IRawMailboxArtifactManager {
+    // @TODO Implement in a follow up PR
+    throw Error('Not implemented');
+  }
+
+  createValidatorAnnounceArtifactManager(
+    _chainMetadata: ChainMetadataForAltVM,
+  ): IRawValidatorAnnounceArtifactManager | null {
+    // @TODO Implement in a follow up PR
+    throw Error('Not implemented');
   }
 
   getMinGas(): MinimumRequiredGasByAction {